+++ src/main/java/model/SimSource.java
@@ -52,7 +52,7 @@ public class SimSource implements ProcessSource{
     public Process generateProcess()
     {
         Random rand = new Random();
-        int memSize = 0;
+        long memSize = 0;
 
         memSize = getRandomIntBetweenRange(10, 30) * 1024;
         Process proc = new Process(procNames.get(rand.nextInt(procNames.size())), ++numOfProcess,
